% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdshinyappar-package.R
\docType{package}
\name{rdshinyappar-package}
\alias{rdshinyappar}
\alias{rdshinyappar-package}
\title{rdshinyappar: Runtime-hjälpare för Region Dalarnas Shiny-appar}
\description{
Hjälpfunktioner som Region Dalarnas Shiny-appar använder i drift: databasuppkopplingar (läs/skriv), lösenordshantering via .Renviron, tabellistning i PostGIS, EWKB-till-sf-konvertering och gemensam shiny.telemetry-loggning. Utbrutet ur func_shinyappar.R i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
