% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losenord.R
\name{shiny_delete_password}
\alias{shiny_delete_password}
\title{Ta bort ett sparat lösenord}
\usage{
shiny_delete_password(service)
}
\arguments{
\item{service}{Tjänstnamn, bara \verb{A-Z a-z 0-9 _}.}
}
\value{
\code{NULL} osynligt.
}
\description{
Tar bort raden \verb{<service>_PWD=...} ur \verb{~/.Renviron}.
}
