% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losenord.R
\name{shiny_get_password}
\alias{shiny_get_password}
\title{Hämta ett sparat lösenord}
\usage{
shiny_get_password(service)
}
\arguments{
\item{service}{Tjänstnamn, bara \verb{A-Z a-z 0-9 _}.}
}
\value{
Lösenordssträngen. Fel om variabeln saknas.
}
\description{
Läser om \verb{~/.Renviron} och returnerar \verb{<service>_PWD}.
}
