% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losenord.R
\name{shiny_list_passwords}
\alias{shiny_list_passwords}
\title{Lista tjänster med sparade lösenord}
\usage{
shiny_list_passwords()
}
\value{
Teckenvektor med tjänstnamn (variabler som slutar på \verb{_PWD} i
\verb{~/.Renviron}, utan suffixet).
}
\description{
Lista tjänster med sparade lösenord
}
