% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetri.R
\name{telemetri_ui}
\alias{telemetri_ui}
\title{Wrapper för \code{use_telemetry()} med NULL-skydd}
\usage{
telemetri_ui(telemetry)
}
\arguments{
\item{telemetry}{Ett Telemetry-objekt från \code{\link[=skapa_telemetry]{skapa_telemetry()}}, eller \code{NULL}.}
}
\value{
UI-taggar från \code{shiny.telemetry::use_telemetry()}, eller \code{NULL}.
}
\description{
Så att \code{ui.R} slipper \code{if}-satsen.
}
